Creates src/components/AuthForm.js.
This commit creates the authform component. This component will
be responsible for signing users up as well as loging them in.
The form can be interacted with and the form currently displays
different input fields based on whatc action is happening. The
onsubmit function for the form still needs to be implemented.