Creates src/store/actions/auth.js.
This commit creates the auth action. This commit implements the
setCurrentUesr action creator which creates the setCurrentUser
action object that is to be dispatched to the reducer. This
commit also implements the authUser function which uses the api
service to call a post request to /api/auth/{type} based on the
type and posts the user data. It then stores the retrieve JWT
token in local storage and then distpachets the setCurrentUser
action with the user data retreived from the back-end.