Updates src/components/AuthForm.js - authUser / button.
This commit upates the authform to include a submit button
on the form and starts the logic for the onsubmit function for
the form. The onsubmit function with use the onAuth prop
(authUser action) and the state to either login in the user
or signup the user based on which form the user is using.