Updating src/containers/app.js - Authorization.
This commit updates the app container to set the authorization
token if one already exists in local storage. It will also
log the user out of the token fails or has been manually tampered
with. This also adds the ability to re-log the user in if
something goes wrong with the app during runtime.