Creates src/hoc/withAuth.js.
This component is a High Order Component or essentially a wrapper
component that has lifecylce methods to verify the user is
authenticated in the redux state. If the user is not
authenticated, the user will be re-routed to the /sigin page.
If the user is authenticated, the component that it is wrapping
is displayed and rendered normally.