Updates src/containers/Main.js - MessageForm.
This commit udpates the Main container to include a route to
create new messages. This route is /user/:id/messages/new. This
is the path where the MessageForm component is rendered. This
component is wrapped with the withAuth component to verify a user
is authenticated before they are allowed to create messages.