Updating src/store/actions/message.js - Remove Messages.
This commit creates the remove message action creator and the
delete message function that handles making the api call to
delete the message for a given user_id and a given message_id.
It then dispatches a remove message action using the remove
message action creator to remove the message from the state.